test cases merged into two classes: tweets & users.
