+++ test/AbstractTest.php
@@ -1,4 +1,4 @@
-<?php
+<?php /** @noinspection PhpIllegalPsrClassPathInspection */
 
 namespace Noweh\TwitterApi\Test;
